tweaks

+++ pkg/game/card.go
@@ -4,7 +4,7 @@ import "fmt"
 
 // Card structure
 type Card struct {
-	Value int8
+	Value int
 }
 
 // NewCard creates Card with value
@@ -12,5 +12,5 @@ func NewCard(value int) (Card, error) {
 	if value < 1 || value > 11 {
 		return Card{}, fmt.Errorf("Card value (%d) must between 1 and 11", value)
 	}
-	return Card{Value: int8(value)}, nil
+	return Card{Value: value}, nil
 }
